+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/users/[userId]/page.tsx
@@ -1,0 +1,11 @@
+import PageClient from "./page-client";
+
+export const metadata = {
+  title: "User Profile",
+};
+
+export default async function Page(props: { params: { userId: string } }) {
+  return (
+    <PageClient userId={props.params.userId} />
+  );
+}
